HTML, CSS and JS.
Only thing missing is for it to work. draw works but when rezised or
deleted it doesn't. something strange happens when the clear function
is added to event listener of the clear button where it triggers without
any clicks. whole thing is messed up and I have made it worse by trying
to fix it.